# Avoid round trips while fixing index names