--- a/src/backend/distributed/sql/udfs/citus_run_local_command/latest.sql
+++ b/src/backend/distributed/sql/udfs/citus_run_local_command/latest.sql
@@ -1,0 +1,8 @@
+CREATE OR REPLACE FUNCTION pg_catalog.citus_run_local_command(command text)
+RETURNS void AS $$
+BEGIN
+    EXECUTE $1;
+END;
+$$ LANGUAGE PLPGSQL;
+COMMENT ON FUNCTION pg_catalog.citus_run_local_command(text)
+    IS 'citus_run_local_command executes the input command';