--- a/src/backend/distributed/utils/multi_partitioning_utils.c
+++ b/src/backend/distributed/utils/multi_partitioning_utils.c
@@ -569,7 +569,23 @@ CreateFixPartitionShardIndexNamesTaskList(Oid parentRelationId, Oid partitionRel
 			task->taskId = taskId++;
 
 			task->taskType = DDL_TASK;
-			SetTaskQueryStringList(task, queryStringList);
+
+			/*
+			 * There could be O(#partitions * #indexes) queries in
+			 * the queryStringList.
+			 *
+			 * In order to avoid round-trips per query in queryStringList,
+			 * we join the string and send as a single command via the UDF.
+			 * Otherwise, the executor sends each command with one
+			 * round-trip.
+			 */
+			char *string = StringJoin(queryStringList, ';');
+			StringInfo commandToRun = makeStringInfo();
+
+			appendStringInfo(commandToRun,
+							 "SELECT pg_catalog.citus_run_local_command($$%s$$)", string);
+			SetTaskQueryString(task, commandToRun->data);
+
 			task->dependentTaskList = NULL;
 			task->replicationModel = REPLICATION_MODEL_INVALID;
 			task->anchorShardId = parentShardId;