--- a/src/test/regress/sql/citus_run_command.sql
+++ b/src/test/regress/sql/citus_run_command.sql
@@ -1,0 +1,4 @@
+SELECT citus_run_local_command($$SELECT 1; SELECT 1$$);
+SELECT citus_run_local_command($$SELECT 1; SELECT 1/0$$);
+SELECT citus_run_local_command(NULL);
+