--- a/src/test/regress/sql_schedule
+++ b/src/test/regress/sql_schedule
@@ -5,4 +5,4 @@ test: ch_benchmarks_4 ch_benchmarks_5 ch_benchmarks_6
 test: intermediate_result_pruning_queries_1 intermediate_result_pruning_queries_2
 test: dropped_columns_1 distributed_planning
 test: local_dist_join
-test: connectivity_checks
+test: connectivity_checks citus_run_command