… src/test/regress/expected/citus_local_table_triggers.out
… src/test/regress/expected/coordinator_shouldhaveshards.out